brewcop.py: add urwid based user interface
This is a first cut at an urwid (curses-like) user interface
for brewcop.  It implements a simple screen layout with buttons
compatible with the raspberry pi touchscreen, at least when
running under the full GUI desktop.  Thus far, it deals with
the scale but does not generate Slack notifications.

Aside: although urwid can also work with the raw linux console,
pi touchscreen "clicks" were ignored when GDM is not installed,
and did not seem to be able to select anything with GDM installed.
For now, run brewcop.py in the GUI desktop in an lxterm window.

A Scale class is implemented for the Avery Berkel scale that
includes lbs to g conversion and a Tare (zero) function.  For now
it calls out to the C program for querying the current reading.
A little indicator appears in the upper right hand corner of the
screen while polling is in progress.  It goes away immediately
on success.  On failure it turns red and puts "----" in the scale
readout window.

A Brewcop class implements the application logic of Brewcop.
This class is not yet filled in.  For now it receives mode
changes as the user selects them via radio buttons, and
weight updates as the scale is polled.  It will need to implement
the logic for Slack notifications when in "Brew" mode, based
on the evolution of scale readings.

The screen is divided up into a one-line header, a main screen
that contains an ASCII art background image overlayed with the
scale readout window, and a footer containing buttons.  The
left-most three buttons are a radio set for selecting the mode:
"Beans" (weiging beans), "Brew" (brewing/monitoring coffee),
and "Clean" (cleaning the machine).  The right-most button is
a "Tare" button for zeroing the scale with an empty vessel on
it (bean cup or empty coffee pot).

Every 2 seconds, the scale is polled.  This is under the control
of urwid's event looop.  While urwid is designed to integrate
with applications that use the twisted reactor, in this iteration,
we just let urwid drive the application, and block the reactor while
the scale is being polled.  This leads to some sluggishness if one
presses a button while the poll is progress.  Eventually we should
look at integrating the scale with the event loop.