brewcop.py: restore original tare method
Problem: The scale "zero" method does not work for
containers as heavy as the carafe so I guess that was
a bad move.

Restore the previous "tare" method, which is implemented
only within brewcop.py, thus the scale's readout is
not zeroed.  Leave the scale "zero" method in place,
in case it's useful in the future.