brewcop.py: add fake scale fallback
Problem: for development, it would be convenient if one could
start brewcop.py without a scale attached.

If Scale class cannot be initialized, use the NoScale class,
which inherits from Scale and overrides the active methods
poll() and zero() with no-ops, and overrides the display
attribute to unconditionally return "no scale".