brewcop.py: re-implement Brewcop class
Move the global stuff into the Brewcop class.  Change the logic
so that there are two modes, offline and online.  Offline displays
the scale readout and a message "brewcop is offline".  Online
displays a progress bar using hardwired pot tare and capacity.

The mode is offline when scale reads below the pot tare value.

This bimodal scheme does not rely on (somewhat broken) input from
the py touchscreen and thus can operate outside of the X windows
environment, directly on the console.

Other changes:
- Add inline documentation for classes and methods
- Add Scale.at_zero() for testing ECR zero status
- Run through black formatter (one small change from before)