brewcop.py: add Brains class
Add a new class that Brewcop feeds pot content
weight to while online.  It retains 30s worth of
history and implements a simple state machine.
The states are:

unknown - history is empty

brewing - history shows content weight increasing.

empty - not brewing, and most recent history value
is under a threshold.

ready - not brewing or empty, and history is
stable/decreasing

Show the state in the center of the top line of
the display, with the time elapsed since the state
was entered.

The ready state has some additional features:
1) If ready is entered from brewing (as opposed to
from unknown), a Slack notificiation stub is called,
currently a no-op.
2) If the elapsed time since ready was entered is
above 4h, change color of the state display from
green to red to let people know at a glance that
coffee is stale.