add back slack notify capability
Problem: slack notifications were dropped in the rewrite.

Add them back, but this time use a webhook instead of the
python SlackClient (now slack_sdk) as it's simpler, and the
interfaces we used in 2018 are now deprecated.

When configuring the brewcop "app" in Slack, select
"Incoming Webhooks", enable them, and select the channel in
your workspace that brewcop should post in.  In brewcop's
.bashrc, set and export SLACK_WEBHOOK_URL to the URL shown
so that brewcop.py will pick it up when it runs.

Restore the messages we thought up during the hackathon, plus
add a special one for Stephen.