allow small variation in technivorm pot weight
Problem: pot_tare_g is hard coded to 796g, the original weight
of the pot at work.  My pot at home reads 794g, so it shows offline
when placed on the scale empty.

Set pot_tare_g to 795 and call the tared weight zero if it is
within +/-4g of zero.